refactor(nodebuilder/node): split out admin into separate file and address ryans comments
